refactor：storybook：component：types移動
