+++ src/stories/components/atoms/Button/FormStateButton/types/index.ts
@@ -1,0 +1,4 @@
+import { ButtonProps } from '@/stories/components/atoms/Button/Base/types';
+export type FormStateButtonProps = ButtonProps & {
+  processing: boolean;
+};
