+++ src/stories/components/atoms/Icon/Base/index.tsx
@@ -1,15 +1,13 @@
-import { SvgIconProps, SvgIcon } from '@mui/material';
-import { ReactNode, FC } from 'react';
+import { SvgIcon } from '@mui/material';
+import { FC } from 'react';
 
-export type CustomSvgIconProps = SvgIconProps & {
-  icon: ReactNode;
-};
+import { CustomSvgIconProps, IconProps } from './types';
+
 const CustomSvgIcon: FC<CustomSvgIconProps> = ({
   icon,
   ...props
 }: CustomSvgIconProps) => <SvgIcon {...props}>{icon}</SvgIcon>;
 
-export type IconProps = CustomSvgIconProps;
 export const Icon: FC<IconProps> = ({ icon, ...props }: IconProps) => (
   <CustomSvgIcon {...props} icon={icon} />
 );
