+++ src/stories/components/atoms/Icon/Base/types/index.ts
@@ -1,0 +1,7 @@
+import { SvgIconProps } from '@mui/material';
+import { ReactNode } from 'react';
+
+export type CustomSvgIconProps = SvgIconProps & {
+  icon: ReactNode;
+};
+export type IconProps = CustomSvgIconProps;
