+++ src/stories/components/atoms/Icon/Basic/index.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react';
 
-import { Icon as BaseIcon, IconProps } from '../Base';
+import { Icon as BaseIcon } from '../Base';
+import { IconProps } from '../Base/types';
 
 export const Icon: FC<IconProps> = ({ icon }: IconProps) => {
   return <BaseIcon icon={icon} fontSize='large' />;
