+++ src/stories/components/atoms/IconButton/Base/types/index.ts
@@ -1,0 +1,7 @@
+import { ButtonProps } from '@/stories/components/atoms/Button/Base/types';
+
+export type IconButtonProps = ButtonProps & {
+  icon?: React.ReactNode;
+  label?: string;
+  disabled?: boolean;
+};
