+++ src/stories/components/atoms/ListItem/Base/types/index.ts
@@ -1,0 +1,7 @@
+import { ListItemProps as MuiListItemProps } from '@mui/material';
+import { MouseEventHandler } from 'react';
+
+export type ListItemProps = MuiListItemProps & {
+  label?: string;
+  selectHandler?: MouseEventHandler<HTMLLIElement>;
+};
