+++ src/stories/components/atoms/Select/Base/types/index.ts
@@ -1,0 +1,5 @@
+import { SelectProps as MuiSelectProps, SelectChangeEvent } from '@mui/material';
+
+export type SelectProps = Omit<MuiSelectProps, 'onChange'> & {
+  selectHandler?: (event: SelectChangeEvent<unknown>) => void;
+};
