+++ src/stories/components/atoms/Select/Basic/index.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react';
 
-import { Select as BaseSelect, SelectProps } from '../Base';
+import { Select as BaseSelect } from '../Base';
+import { SelectProps } from '../Base/types';
 
 export const Select: FC<SelectProps> = ({ selectHandler }: SelectProps) => {
   return <BaseSelect selectHandler={selectHandler} />;
