+++ src/stories/components/atoms/TextField/Base/types/index.ts
@@ -1,0 +1,9 @@
+import { TextFieldProps as MuiTextFieldProps } from '@mui/material';
+import { ChangeEventHandler } from 'react';
+
+export type TextFieldProps = Omit<MuiTextFieldProps, 'onChange'> & {
+  inputHandler?: ChangeEventHandler<HTMLInputElement> | undefined;
+  disabledFlg?: boolean;
+  requiredFlg?: boolean;
+  value?: string;
+};
