+++ src/stories/components/atoms/TextField/Basic/types/index.ts
@@ -1,0 +1,5 @@
+import { TextFieldProps as BaseTextFieldProps } from '../../Base/types';
+
+export type TextFieldProps = BaseTextFieldProps & {
+  resetTextValue?: string;
+};
