+++ src/stories/components/molecules/Button/LogoutButtonFlex/index.tsx
@@ -2,15 +2,13 @@ import EditNoteIcon from '@mui/icons-material/EditNote';
 import PersonAddIcon from '@mui/icons-material/PersonAdd';
 import { FC, useContext } from 'react';
 
+import { LogoutButtonFlexProps } from './types';
+
 import { SetUserDataContext } from '@/stories/common/context';
 import { FlexBox } from '@/stories/components/atoms/Box/FlexBox';
 import { IconButton } from '@/stories/components/atoms/IconButton/Base';
 import { LogoutButton } from '@/stories/components/atoms/IconButton/LogoutButton';
 import { RegisterButton } from '@/stories/components/atoms/IconButton/RegisterButton';
-
-export type LogoutButtonFlexProps = {
-  dataExistFlg: boolean;
-};
 
 export const LogoutButtonFlex: FC<LogoutButtonFlexProps> = ({
   dataExistFlg,
