+++ src/stories/components/molecules/Button/LogoutButtonFlex/types/index.ts
@@ -1,0 +1,3 @@
+export type LogoutButtonFlexProps = {
+  dataExistFlg: boolean;
+};
