+++ src/stories/components/molecules/Button/SendStateButton/types/index.ts
@@ -1,0 +1,6 @@
+import { ButtonProps } from '@/stories/components/atoms/Button/Base/types';
+
+export type SendStateButtonType = ButtonProps & {
+  keepHandler: () => void;
+  sendHandler: () => void;
+};
