+++ src/stories/components/molecules/List/LogsList/index.tsx
@@ -1,13 +1,9 @@
 import { FC } from 'react';
 
+import { LogsListProp } from './types';
 import { LogsListItem } from '../../ListItem/LogsListItem';
 
-import { MessageType } from '@/stories/common/types/db';
 import { List } from '@/stories/components/atoms/List/Base';
-
-export type LogsListProp = {
-  messageLog: MessageType[];
-};
 
 export const LogsList: FC<LogsListProp> = ({ messageLog }: LogsListProp) => {
   return (
