+++ src/stories/components/molecules/List/LogsList/types/index.ts
@@ -1,0 +1,5 @@
+import { MessageType } from '@/stories/common/types/db';
+
+export type LogsListProp = {
+  messageLog: MessageType[];
+};
