+++ src/stories/components/molecules/List/PersonList/types/index.ts
@@ -1,0 +1,4 @@
+import { MouseEventHandler } from 'react';
+export type PersonListProps = {
+  selectHandler?: MouseEventHandler<HTMLLIElement>;
+};
