+++ src/stories/components/molecules/ListItem/LogsListItem/types/index.ts
@@ -1,0 +1,6 @@
+import { MessageType } from '@/stories/common/types/db';
+import { ListItemProps } from '@/stories/components/atoms/ListItem/Base/types';
+
+export type LogsListItemProps = ListItemProps & {
+  item: MessageType;
+};
