+++ src/stories/components/molecules/ListItem/PersonListItem/types/index.ts
@@ -1,0 +1,6 @@
+import { ListItemProps } from '@/stories/components/atoms/ListItem/Base/types';
+
+export type PersonListItemProps = ListItemProps & {
+  id: string;
+  message: string;
+};
