+++ src/stories/components/molecules/Title/types/index.ts
@@ -1,0 +1,3 @@
+export type TitleProps = {
+  title: string;
+};
