+++ src/stories/components/templates/CommonTemplate/types/index.ts
@@ -1,0 +1,7 @@
+import { TitleProps } from '@/stories/components/molecules/Title/types';
+
+export type CommonTemplateProps = TitleProps;
+export type GetComponentProps = CommonTemplateProps & {
+  padding: number;
+  component: React.ReactNode;
+};
